+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the "QuoteWithMarker" extension will be documented in this file.
 
+## [1.1.0]
+
+### Leading zeros
+
+Added leading zeros to month and day to always get the same date format.
+
+### blockComment
+
+Some languages do not have line comments, such as CSS. However, they have the possibility to comment out a code block (blockComment).
+Now the QuoteWithMarker can be used in these languages as well.
+
 ## [1.0.3]
 
 ### Maintenance